requirements.txt
ccxt
yfinance
pandas
ta
requests
apscheduler
anthropic
python-dotenv
google-genai